requirements.txt
openai==1.52.2
sib-api-v3-sdk==7.6.0
requests==2.32.3
praw==7.8.1
python-dotenv